fix dependencies for windows
