+++ siquery_tables/src/linux/groups.rs
@@ -1,5 +1,4 @@
 use libc::{
-    int32_t,
     getgrent,
     setgrent,
     endgrent,
