+++ siquery_tables/src/utils.rs
@@ -1,5 +1,3 @@
-use libc;
-
 /// Remove trailing '\n' at the end of a string.
 pub fn trim_string(s: &mut String) {
     while s.ends_with('\n') || s.ends_with('\r') || s.ends_with(',') {
